% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{availableLangs}
\alias{availableLangs}
\title{Display available languages}
\usage{
availableLangs(localeDir = NULL)
}
\description{
Call this function to display the language code for languages currently
supported by \code{shi18ny}.
}
