% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{i_}
\alias{i_}
\title{Translate text}
\usage{
i_(str, lang = NULL, i18n = NULL, markdown = FALSE, keys = c("name", "label"))
}
\arguments{
\item{str}{A string, vector or list to be translated.}

\item{lang}{Code for the language that the original language should be
translated into.}

\item{i18n}{List of language configurations. Options that can be set
are:

\code{defaultLang} Default language used in Shiny app

\code{availableLangs} Language that can be chosen for translation in Shiny app;
there are currently 15 available languages

\code{localeDir} Directory to \code{yaml} files which contain custom keyword
translations

\code{fallbacks} List of fallback languages if translation for a word is not
found in desired language

\code{queryParameter} String to define query parameter if language to be set
through URL}

\item{markdown}{TODO}

\item{keys}{If \code{str} is a list this is a string (or a vector of strings)
specifying which key(s) of the list to translate.}
}
\value{
Translation of input text in the same format as the input.
}
\description{
Translate strings, vectors or lists in the Server of a Shiny app.
}
\details{
There are currently 15 languages available for translation:\tabular{ll}{
   code \tab language \cr
   ar \tab Arabic \cr
   ca \tab Catalan \cr
   da \tab Danish \cr
   de \tab German \cr
   en \tab English \cr
   es \tab Spanish \cr
   fr \tab French \cr
   he \tab Hebrew \cr
   hi \tab Hindi \cr
   it \tab Italian \cr
   pt \tab Portuguese \cr
   pt_BR \tab Portuguese (Brazil) \cr
   ru \tab Russian \cr
   sv \tab Swedish \cr
   zh_CN \tab Chinese \cr
}
}
\examples{
i_("hello", lang = "de")

i_(c("hello", "world"), lang = "de")

i_(list(id = "hello", translate = "world"), lang = "es", keys = "translate")
}
