% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{langSelector}
\alias{langSelector}
\title{Language selector}
\usage{
langSelector(input, output, session, i18n = NULL, showSelector = TRUE)
}
\arguments{
\item{i18n}{List of language configurations; can only be set for \code{i_}.
Options that can be set are:

\code{defaultLang} Default language used in Shiny app; default = "en"

\code{availableLangs} Language that can be chosen for translation in Shiny app;
there are currently 15 \strong{Available languages} (see below);
defaults to all.

\code{localeDir} Directory to \code{yaml} files which contain custom keyword
translations; default = "locale"

\code{fallbacks} List of fallback languages if translation for a word is not
found in desired language; defaults to \strong{Default fallbacks} (see below)

\code{queryParameter} String to define query parameter if language to be set
through URL; default = "lang"}

\item{showSelector}{Boolean to specify if language selector should be
displayed or not. Even if selector is not displayed, language can always
be set by query parameter; default = "lang"}
}
\value{
String of language code of currently selected languages.
}
\description{
This Shiny module gets language that's currently selected in the language
selector. It needs to be used in the Server to save the currently active
(selected) language in a reactive. This reactive can then be passed to the
\code{\link{i_}} function to translate the text.
}
\section{Available languages}{
 There are currently 15 languages available for
translation:\tabular{ll}{
   code \tab language \cr
   ar \tab Arabic \cr
   ca \tab Catalan \cr
   da \tab Danish \cr
   de \tab German \cr
   en \tab English \cr
   es \tab Spanish \cr
   fr \tab French \cr
   he \tab Hebrew \cr
   hi \tab Hindi \cr
   it \tab Italian \cr
   pt \tab Portuguese \cr
   pt_BR \tab Portuguese (Brazil) \cr
   ru \tab Russian \cr
   sv \tab Swedish \cr
   zh_CN \tab Chinese \cr
}
}

\section{Default fallbacks}{
 If no fallback languages are specified,
translations automatically fall back onto the following languages.\tabular{ll}{
   original language \tab fallback language \cr
   es \tab pt \cr
   pt \tab es \cr
   fr \tab pt \cr
   de \tab nl \cr
   nl \tab de \cr
}
}

\examples{
\dontrun{
i18n <- list(
  defaultLang = "en",
  availableLangs = c("en", "de")
)
lang <- callModule(langSelector,"lang", i18n = i18n, showSelector = TRUE)
}
}
