% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{langSelectorInput}
\alias{langSelectorInput}
\title{Language selector input}
\usage{
langSelectorInput(id, position = "right", width = 80)
}
\arguments{
\item{id}{The id that will be used to access the value.}

\item{position}{String indicating position of language switcher; options
are "right" and "fixed"; default = "right".}

\item{width}{Width of language switcher container; default = 80.}
}
\description{
This function needs to be included in the UI to create the
language switcher in the Shiny app.
}
