% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExampleFiles.R
\name{runExample}
\alias{runExample}
\title{Run one of the \code{shi18ny} examples.}
\usage{
runExample(example_name = NULL)
}
\arguments{
\item{example_name}{Name of the example}
}
\description{
For a list of available examples, run \code{shi18ny::runExample()}.
}
