% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectLangInput.R
\name{selectLangInput}
\alias{selectLangInput}
\title{Add images to dropdown options}
\usage{
selectLangInput(
  inputId,
  label,
  langs,
  flags = NULL,
  show_flags = TRUE,
  selected = NULL,
  placeholder = NULL,
  width = 100
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{placeholder}{HTML to render as placeholder, defaults to empty div.}

\item{width}{width in of input.}

\item{choices}{List of values to select from, when named the names are
appended to the right of the image.}

\item{images}{List of image location that can be put in a src attribute.}
}
\description{
This function works only with bootstrap for now. Borrowed from
the package shinyinvoer.
}
