% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{uiLangUpdate}
\alias{uiLangUpdate}
\title{Update language in UI translations}
\usage{
uiLangUpdate(classes, lang)
}
\arguments{
\item{classes}{UI classes; pass \code{input$shi18ny_ui_classes} to refer to all
instances of \code{\link{ui_}}.}

\item{lang}{Code for the language that the original language should be
translated into.

In the following example, \code{lang()} is the reactive value created by the
\code{\link{langSelector}} function.}
}
\description{
This function needs to be included in the Server of the Shiny app if if text
is translated in the UI of the app using the \code{\link{ui_}} function. It
makes sure the language used by \code{\link{ui_}} to do the translation is
updated when the active (selected) language changes.
}
\examples{
\dontrun{
observeEvent(lang(),{
  uiLangUpdate(classes = input$shi18ny_ui_classes,
               lang = lang())
  })
}
}
