% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-selectImage.R
\name{updateSelectImageInput}
\alias{updateSelectImageInput}
\title{Update select image input}
\usage{
updateSelectImageInput(
  session,
  inputId,
  label = NULL,
  choices = NULL,
  images = NULL,
  selected = NULL
)
}
\arguments{
\item{session}{Shiny session}

\item{inputId}{The input slot that will be used to access the value.}

\item{choices}{List of values to select from, when named the names are
appended to the right of the image.}

\item{images}{List of image location that can be put in a src attribute.}

\item{selected}{Selected image, defaults to first one.}

\item{placeholder}{HTML to render as placeholder, overrides selected param.}

\item{width}{width in of input.}
}
\description{
Update select image input
}
