% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{useShi18ny}
\alias{useShi18ny}
\title{Initialize shi18ny}
\usage{
useShi18ny()
}
\description{
This function needs to be included in the UI of the Shiny app. It runs the
required client side javascript code to update strings when the currently
active (selected) language changes.
}
